First dedicated version for Open Vision, no more support for other images:
Well I tried to have a common ground with other teams but with no luck: https://forums.openpli.org/topic/77494-one-proc-file-for-detecting-the-machine-in-all-enigma2-images/

We check things better and faster in OV and it's not my duty to support all images so sorry and EOL for master branch.

Any developer could ask me for write access to Backup Suite and could maintenance the master branch but I will work on develop branch for OV since now.

As always PRs are welcome.

Good luck.